Handle inputs larger the Integer.MAX_VALUE for review limit

+++ src/test/java/seedu/flashnotes/logic/commands/DeleteDeckCommandTest.java
@@ -233,12 +233,12 @@ public class DeleteDeckCommandTest {
         }
 
         @Override
-        public Integer getReviewCardLimit() {
+        public long getReviewCardLimit() {
             throw new AssertionError("This method should not be called.");
         }
 
         @Override
-        public void setReviewCardLimit(Integer reviewCardLimit) {
+        public void setReviewCardLimit(long reviewCardLimit) {
             throw new AssertionError("This method should not be called.");
         }
